Fix edge case bug in merge join operator (#494)
An update to a row on the right side edge of a page with a larger value in the other columns would add it to the next page. Then the delete of the old value would mean that the key points to a page where the key does not exist. This change allows the iterator to check the next page for the key. This is only useful really when searching for a single column while insertion is done based on multiple columns.