Add initial version of the delta lake sink (#692)
* Start on delta lake sink

This commit only contains code for parsing statistics and comparing schema

* add writing

* continue on sink

statistics remain when writing and also when checking existing data to skip files

* change how statistics is parsed and written.

Now files are written with correct statistics, next step is data skipping files when finding rows to remove.

* Add data file skipping based on statistics

* move write delete files to own method

* Add timestamp statistics comparer

* Add more writers for types, added struct support

More tests are required, and array and map type. Not possible right now to create a new table as a struct,map or list.
This should be added before release so one can configure the destination schema in sql

* start on struct statistics

* add struct statistics

* add code to write new files on delete if delete percentage is too high

* Add tests for all implemented data types to write, missing array and map

* Start on array writer

Must implement index and count on all copy array methods to be able to skip values

* Add array type support

* Add comparers for the different types

* change to use new comparers to find rows in file

* Add support for map, also fix bug in copy array

* fix struct statistics comparer

* add more tests for creating a table if it does not exist

This involves adding support for more data types in sql to more clearly define the rules.

* Add support to create a map type column in sql with delta

* add start on docs

* add support to write cdc files

* Add more docs, add partition writers, but did not add code to use them

* remove commented code and fix warnings

* fix failing tests

* fix so change data is enabled based on the configuration of the table

* Add config to allow the user to decide if deletion vector should be enabled or not

Also updated the docs to describe the different options