Add string_split function for substring manipulation (#727)
* Add string_split function for substring manipulation

Introduce a new `string_split` function that splits a string into a collection of substrings based on a specified delimiter. If the delimiter is null, the original string is returned as the only element. The function is implemented in `BuiltInStringFunctions.cs` and registered in the `FunctionsString` class. Tests for `string_split` are added in `StringFunctionTests.cs`, along with a new test file `string_split.test` to validate its functionality.

* changed string split to return null if the split char is not a string and not null

* added comment about non-string split delimiter character

* fixed formatting of sql string

* removed string split implementation

* Add StringSplitImplementation to BuiltInStringFunctions

Introduces a new method `StringSplitImplementation` that splits a string into a list based on a specified delimiter. The method handles null values and type checks for both the input string and the delimiter. It is registered as a scalar method in the `functionsRegister`, enhancing string operation capabilities.