Add support for options data source from configuration
This adds a new data source that uses IOptionsMonitor to read updates and outputs the data as a table into the stream.

This change also includes a small fix to include the inner data type from lists during UNNEST since UNNEST would be quite common with this case if a list is used.