Add a new event method before checkpoint is saved for operators
One use case is if a source has an offset (such as Kafka), and only a subset of events are in the result set. If no new event has been sent into the stream after the checkpoint event, this method can be used to update the offset to the latest one to skip loading in unnecessary data in the case of a crash.